synthethic_data/requirements.txt
google-generativeai # For using Gemini API directly
python-dotenv